BGDIINF_SB-2121: Caching topics the same manner as caching translations
