+++ app/__init__.py
@@ -3,7 +3,6 @@ import os
 import re
 import time
 
-import psycopg2 as psy
 from flask_caching import Cache
 from werkzeug.exceptions import HTTPException
 
@@ -13,7 +12,6 @@ from flask import g
 from flask import request
 
 from app import settings
-from app.helpers.utils import get_topics_from_db
 from app.helpers.utils import make_error_msg
 
 logger = logging.getLogger(__name__)
@@ -25,18 +23,6 @@ app = Flask(__name__)
 app.config.from_object(settings)
 
 cache = Cache(app)
-
-# TODO: This will have to be discussed in a general manner.
-# Right now the topics do not serve anything else than returning HTTP 400
-# when the topic does not exist. No filtering is being done on it at all
-# f.ex searching for lagefixpunkte in topic bafu, which is meaningless
-# https://api3.geo.admin.ch/2111231107/rest/services/bafu/SearchServer?sr=2056&searchText=CH030000123112311020&lang=en&type=featuresearch&features=ch.swisstopo.fixpunkte-lfp1&timeEnabled=false&timeStamps=
-try:
-    topics = get_topics_from_db()
-except psy.Error:
-    topics = settings.FALLBACK_TOPICS
-    logger.error("Connection to the db base could not be established." \
-            " Using fallback %s", settings.FALLBACK_TOPICS)
 
 
 # NOTE it is better to have this method registered first (before validate_origin) otherwise
