+++ app/helpers/utils.py
@@ -2,13 +2,11 @@ import logging
 import logging.config
 from os import path
 
-import psycopg2 as psy
 import yaml
 
 from flask import jsonify
 from flask import make_response
 
-from app import settings
 from app.settings import ALLOWED_DOMAINS
 from app.settings import LOGGING_CFG
 from app.settings import LOGS_DIR
@@ -37,52 +35,3 @@ def get_logging_cfg():
 def init_logging():
     config = get_logging_cfg()
     logging.config.dictConfig(config)
-
-
-def get_topics_from_db():
-    '''Get a list with all topics from bod
-
-    Returns:
-        A List with the topics or an empty list
-    '''
-    logger.debug('Connecting to %s db on host %s', settings.BOD_DB_NAME, settings.BOD_DB_HOST)
-    try:
-        connection = psy.connect(
-            dbname=settings.BOD_DB_NAME,
-            user=settings.BOD_DB_USER,
-            password=settings.BOD_DB_PASSWD,
-            host=settings.BOD_DB_HOST,
-            port=settings.BOD_DB_PORT,
-            connect_timeout=5
-        )
-    except psy.Error as error:
-        logger.error("Unable to connect: %s", error)
-        if error.pgerror:
-            logger.error('pgerror: %s', error.pgerror)
-        if error.diag.message_detail:
-            logger.error('message detail: %s', error.diag.message_detail)
-        raise
-
-    # Open cursor for DB-Operations
-    cursor = connection.cursor()
-
-    try:
-        # select records from DB
-        cursor.execute("""
-            SELECT topic FROM "re3".topics
-            """)
-    except psy.Error as error:
-        logger.exception('Failed to retrieve wms config from DB: %s', error)
-        raise
-
-    total_records = cursor.rowcount
-    logger.info("Found %s records", total_records)
-
-    # iterate through table
-    _topics = ['all']
-    for i, record in enumerate(cursor):
-        logger.debug('topic in topics %d: %s', i, record)
-        _topics.append(record[0])
-
-    logger.info("List of topics has been generated")
-    return _topics
